Change point echo to all-client broadcast & add types.
This update adds types for Dialyzer (something missing entirely from
the original version), comment explanations, and references to
the Erlang docs for newcomers reading the code and wanting to base
their own network thingy on this one.

The point echo has also been changed to an all-client echo, so
now this is actually a super naive telnet single-channel chat
server (one that lacks every feature one might ever want, except
for it being super fast and very hard to crash).